refactor: extract network volume diagnostics to separate module
- move network volume diagnostic functions from handler.py to src/network_volume.py
- remove obsolete bash diagnostics from src/start.sh (replaced by python diagnostics)
- update handler.py to import diagnostics from network_volume module
- update dockerfile to include src/network_volume.py in image build
- reduces handler.py size by ~150 lines for better maintainability
